+++ packages/expo-modules-autolinking/src/reactNativeConfig/reactNativeConfig.ts
@@ -20,6 +20,9 @@ import type {
   RNConfigReactNativeProjectConfig,
   RNConfigResult,
 } from './reactNativeConfig.types';
+import { resolveGradlePropertyAsync } from '../platforms/android';
+
+const EDGE_TO_EDGE_ENABLED_GRADLE_PROPERTY_KEY = 'expo.edgeToEdgeEnabled';
 
 /**
  * Create config for react-native core autolinking.
@@ -36,9 +39,14 @@ export async function createReactNativeConfigAsync({
   };
 
   // For Expo SDK 53 onwards, `react-native-edge-to-edge` is a transitive dependency of every expo project. Unless the user
-  // has also included it as a project dependency, we have to autolink it (transitive non-expo module dependencies are not autolinked).
+  // has also included it as a project dependency, we have to autolink it manually (transitive non-expo module dependencies are not autolinked).
+  // There are two reasons why we don't want to autolink `edge-to-edge` when `edgeToEdge` property is set to `false`:
+  // 1. `react-native-is-edge-to-edge` tries to check if the `edge-to-edge` turbomodule is present to determine whether edge-to-edge is enabled.
+  // 2. `react-native-edge-to-edge` applies edge-to-edge in `onHostResume` and has no property to disable this behavior.
   const shouldAutolinkEdgeToEdge =
-    platform === 'android' && !('react-native-edge-to-edge' in dependencyRoots);
+    platform === 'android' &&
+    (await resolveGradleEdgeToEdgeEnabled(projectRoot)) &&
+    !('react-native-edge-to-edge' in dependencyRoots);
 
   if (shouldAutolinkEdgeToEdge) {
     const edgeToEdgeRoot = resolveEdgeToEdgeDependencyRoot(projectRoot);
@@ -220,4 +228,16 @@ export async function resolveAppProjectConfigAsync(
   }
 
   return {};
+}
+
+/**
+ * Resolve the `expo.edgeToEdgeEnabled` property from the `gradle.properties` file.
+ */
+async function resolveGradleEdgeToEdgeEnabled(projectRoot: string): Promise<boolean> {
+  return (
+    (await resolveGradlePropertyAsync(
+      path.join(projectRoot, 'android'),
+      EDGE_TO_EDGE_ENABLED_GRADLE_PROPERTY_KEY
+    )) === 'true'
+  );
 }